refactor:useTableFilter hook 생성
